container_lambda_to_run_rag_slm/docker/requirements.txt
boto3
llama-cpp-python
chromadb
langchain
langchain-chroma
gpt4all
langchainhub
langchain-community
huggingface-hub 
sentence_transformers 
BeautifulSoup4